fastapi/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
httpx==0.28.1
sqlalchemy==2.0.49
pymysql==1.1.2
python-dotenv==1.2.1
apscheduler==3.10.4
playwright==1.44.0
